更新使用spring quartz定时任务实现热帖排行，使用分数形式评价帖子是否为热帖，并采用quartz中的定时任务每5分钟刷新一次

+++ src/main/java/com/sherlockmen/newcoderBlog/entity/DiscussPost.java
@@ -105,7 +105,7 @@ public class DiscussPost {
         return score;
     }
 
-    public void setScore(int score) {
+    public void setScore(double score) {
         this.score = score;
     }
